Remove O major version restriction from shared example
MSP-9923

When I ported version_spec.rb to be a shared example, I forgot to remove
the example that checked that MAJOR was 0.  The 0 check should only be
used for unstable APIs and the shared examples are supposed to work for
all semantic versions, so it should be removed from the shared example.